ILinePlot: don't call data_frame what's just a python dict
Since the dawn of time we have a "data_frame" variable in ILinePlot that
is actually defined as a pd.Series() but it's acting as a dict.  Rename
it and turn the pd.Series into a dict, as it is simpler and it really is
what this variable moves around.