libs/utils/energy: add support for user-defined configurations
The configuration of an energy meter can requires the specification of
a custom set of parameters. For example, using the ARM EnergyProbe (AEP)
we may want to specify the values of a shunt resistor or which channels
to use.

This patch allows to specify an energy meter configuration, for example:

    "emeter" : {
        "instrument" : "aep",
        "conf" : {
            'labels'          : ['LITTLE', 'BIG'],
            'resistor_values' : [0.099, 0.033],
            'device_entry'    : '/dev/ttyACM0',
        }

Which configures two channels with different shunt values.

If an "emeter" configuration section is not specified by the user, the default
one is used, when hardcoded for the specifed board.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>