libs/utils/energy: AEP: add multi-channel support
The ARM EnergyProbe (AEP) allows to monitor up to 3 channels but the current
integration in LISA allows to sample only one channel.

This patch removes that limitation and it allows to sample and report
energy for all the available channels which have been configured by the
user.

A new EnergyCounter named tuple is used to keep track of the power and energy
values computed for each channel. Moreover, the collected CSV file containing
all the collected samples is saved in the results folder, named "samples.csv".
This file can be used for example to plot and/or post process the collected
energy values.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>