ILinePlot: preserve order in the legend of ILinePlot
In aa4bece3adc7 ("ILinePlot: don't call data_frame what's just a
python dict") we turned a pd.Series to a dict, because we were
effectively using it as a dict.  However, there was a behaviour change
as a result of this:  pandas series preserve the order of entries
whereas python dictionaries may reorder its entries.  Use an OrderedDict
instead, as that's what aa4bece3adc7 should have done.