libs/utils/android: Add Screen control class
When working on an Android target, some experiments requires to properly
setup some components, especially to reduce variations on energy
consumptions. The screen is one of the most energy consuming componets.

This patch adds a Screen module which exposes some of the most common
operations to tune the Screen settings of an Android target.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>