libs/utils/android: Add Workload base class
Some workloads can be easily automated for an Android target by just
issuing a proper set of ADB commands. This patch adds a base class to
run workloads on Android.

The goal of this base class is to make it simple to add new ADB-based
automation to run specific workloads on Android. An Android workload is
required to extend this class with a new one which is defined in a
single Python source located under:
   libs/utils/android/workloads

The base class is defined to list all the python files available under
that folder and check if the (optionally) required package is installed
in the target.

The base class provides also a factory method which allows to get an
instance of a specified workload, if available on the target.
For example, a basic usage of this class is:

wload = Workload.get(test_env,
		     wload_name="YouTube")
if wload:
	wload.run(test_env.res_dir,
		  video_url="https://youtu.be/XSGBVzeBUbk?t=45s",
		  video_duration_s=15)

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>