libs/utils/android: Add Jankbench workload
Jankbench is an Android Benchmark application to evaluate frame janks
using many different kind of "scroll-based" workloads.
This patch adds an example of class which inherit from the base
android::Workload class to properly run a Junkbench workload on a
target.

The workload execution provides ADB support to properly start and stop
energy measurements across the execution of the specified test.
Energy samples are collected only if the "energy" tag is found in the
"collect" input string and an energy meter (emeter) has been configured
and properly initialized in the current test environment (test_env).

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>