libs/utils/android: Add YouTube workload
YouTube is a standard Android application which is particularly useful
to evaluate the energy efficiency of an Android target configuration by
just playing a video of a fixed amount of time.

This patch adds an example of class which inherit from the base
android::Workload class to properly playback a YouTube video on a
target.

The workload execution provides ADB support to properly collect frame
statistics after the completion of the video which can be used to
evaluation the performance of the youtube playback.
Energy samples are collected only if the "energy" tag is found in the
"collect" input string and an energy meter (emeter) has been configured
and properly initialized in the current test environment (test_env).

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>