libs/utils/android: Add example notebook with multiple workloads
The android:Workload base class add an abstraction which makes easy to
run a predefined set of workloads on an Android target.

This patch adds a simple example of a notebook which allows to run
YouTube and Janknbench (if available on the target) while collecting
both performance and energy consumption data.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>