libs/utils/trace_analysis: allows to plot tasks on demand
If a set of tasks is not specified when the TraceAnalysis object is created
the plotTask refuses to honour its own "tasks" input parameter.

Let plotTask specify a custom and different set of tasks to be plotted.
This can be useful to exploit Filters functions output to get a list of
tasks of interest to be plotted.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>