libs/utils/trace_analysis: warn on missing data
Let's generate at least a warning when no frequency events are present
in a trace. This usually means that we are using a fixed frequency governor
and the trace has been collected without fake cpu_frequency events injection
at the begin and end of the trace (e.g. using systrace/atrace)

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>