libs/utils/trace: compute time ranges for plot window
If the user does not want the time to be normalized we need to compute the plot
window range boundaries.

Also, the user could specify a window tuple with start and end times. These
values are relative to the basetime of the trace and therefore the plot window
boundaries are computed accordingly.

Signed-off-by: Michele Di Giorgio <michele.digiorgio@arm.com>