ipynb: energy_model: add example for energy model analysis
Add ipynb script to analyse energy model and display the power
efficiency and capacity between two clusters.

This will be helpful to check if energy model is reasonable or not,
and it's straightforward to get to know if there have overlap between
two cluster's energy model; If have overlap, which it's likely to
introduce wrong behavior for EAS energy calculation.

Signed-off-by: Leo Yan <leo.yan@linaro.org>