libs/utils/trace_analysis: add per-CPU frequency residency plots
Generate horizontal bar plots for per-CPU frequency residency.

Frequency residency is the amount of time spent at a each frequency by a given
CPU.

By default all CPUs are taken into account. For each CPU both total and active
(i.e. non-idle) times are plot on the same graph, such that the user has a clear
view of the total time spent at a given frequency and the amount of time in
which the CPU was executing instructions (not idle).

Signed-off-by: Michele Di Giorgio <michele.digiorgio@arm.com>