libs/utils/trace_analysis: fix plotTasks to report one plot per PID
The current version of plot tasks is broken in case a task name matches
multiple PIDs. In that case a plot:
1. report in the title just the first PID
2. mix data for all PIDs in the same plot

This patch fixes this issue by actually producing a set of plots for each
different PIDs which match a specified task name.
Moreover, it is now possible to define the list of tasks to plot by specifying
either task names and/or PIDs.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>