libs/utils/trace_analysis: add plots for per-CPU and per-cluster frequency residency
Two plot modes are available thanks to the `pct` boolean argument.

By default, `pct` is set to False, so total and active residencies are plotted
in terms of time spent at a given frequency.

By setting `pct` to True, total or active residencies for each frequency are
plotted as a percentage relative to the sum of the time intervals computed by
the getFrequencyResidency() functions. For example, suppose total_time is:

    In  [1]: total_time
    Out [1]:
                        time
             frequency
             450        20
             800        45
             950        35

The sum of the time intervals is 100 and we would get a plot showing that 20% of
the time is spent at 450 MHz, 45% at 800 MHz and 35% at 950 MHz.

In case of percentage plots, it is possible to plot either ACTIVE or TOTAL
residencies by properly setting the `active` argument.

Signed-off-by: Michele Di Giorgio <michele.digiorgio@arm.com>