frequency_analysis: use only relevant data for residency computations
When computing frequency residency it is enough to know time stamps and
frequencies. Therefore, it is possible to limit the data to those two
information only.

In this way cluster_freqs becomes a series rather than a data frame, so that all
subsequent operations are all performed on series.

Same applies to the cpu_active signals. The only information needed are time
stamps and cpu idle states.

Signed-off-by: Michele Di Giorgio <michele.digiorgio@arm.com>