libs/utils/analysis: crop data used for analysis according to X range
The user can specify a time range to limit the scope of the trace analysis by
means of the function setXTimeRange().

This patch provides a private method to crop a pandas Series using the specified
range (which by default corresponds to entire duration of the trace).

In order to allow this, the memoized decorator must be removed from
getClusterFrequencyResidency().

Signed-off-by: Michele Di Giorgio <michele.digiorgio@arm.com>