libs/wlgen/rtapp: fix calibration for versions of Android >=N
Apparently, since Android N, by default when running a command via ADB
the STDERR is not forwarded along with the STDOUT and thus the RTApp
calibration fails because we are not able to parse the pLoad value.

This patch force every execution of RTApp to append the STDERR on STDOUT
thus granting to collect from the host side all the output generated by
the application.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>