libs/utils/android: add wrapper for workload-automation
Workload-automation (WA) already provides many workloads for Android targets.
This is a wrapper for LISA to use WA.

In order to use one of the WA workloads (`wa list workloads` tells you what is
available) you need to implement a <workload__class_name> class in a
<workload_name>.py file stored under libs/utils/android/workloads.

<workload_class_name> inherits Wlauto and has to implement the methods from
Workload and (eventually) Wlauto.

WA agendas shall be stored under workloads/agendas.

This way, the workload is visible by the Workload class that will list it as
available and will eventually check that the related package is installed on the
Android target.

Signed-off-by: Michele Di Giorgio <michele.digiorgio@arm.com>