bart: set the version using an explicit file
There is no pythonic way of specifying the version of a project, [0]
describes 7 (seven!) ways of doing it.  We were currently using method
5, setting the value in setup.py and using pkg_resources to get it from
the installed version. This works ok if you have installed the package
using python setup.py or pip, but fails if you are importing bart from a
checkout, which is what lisa do. Even worse, if you import it from lisa
but have an old bart version installed, bart.__version__ will tell you
the version of the installed bart, not the one you have imported and are
using.

Switch to use a version.py file that's distributed with the
project (method 3), as trappy did in ARM-software/trappy@712e6f93b308 .

Fixes #57

[0] https://packaging.python.org/en/latest/single_source_version/