libs/utils/trace: add support to parse functions profiling data
Since this commit:
  082a82c7c ftrace: add support to report function profiling data
devlib provides support to profile a list of specified kernel functions
by specifying them in the TestEnv configurations, e.g.

    my_conf = {
        #... other settings ...
        "ftrace" : {
            "functions" : [
                "select_task_rq_fair",
                "pick_next_task_fair",
            ],
        },
    }

With such a configuration, functions profiling data can be collected
using this API:

    # Collect and keep track of the performance data
    stats_file = os.path.join(te.res_dir, 'trace.stats')
    te.ftrace.get_stats(stats_file)

This patch adds the required support from the LISA side to parse the
functions profiling data and return a data frame for a specified list
of functions.

With this new feature it is now possible to use the Trace object even when
a trace has not been collected. In this case only functions profiling data
frames will be available.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>