libs/utils/analysis: add base support to register trace analysis modules
A TraceModule is a LISA module which can:
a) expose new DataFrame getters
b) expose analysis functions to report/plot DataFrame data

This patch adds the basic support to add new TraceAnalysis modules which
can now be added by simply:
1) creating a new module under libs/utils/analysis
2) derive the top level class from AnalysisModule
3) prefix with "_dfg_" all its methods which generated a DataFrame

The AnalysisRegister class provides the run-time support to register
all the available trace analysis module, which are available as attributes
of this class with a name derived from the module filename.

For example:
   libs/utils/analysis/frequency_analysis.py
will be registered as:
   AnalysisRegister::frequency

The Trace module initialize and keep track of the AnalysisRegister in its
"analysis" attribute. Thus for example the previous module can be accessed
using:

   trace = Trace(platform, trace_dir)
   trace.analysis.frequency.plotClusterFrequencies()

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>