libs/utils/trace: Add min_cluster_cap to task utilization DataFrame
The minimum cluster utilization columns reports the minimum of the maximum
capacities of a cluster which can host the task utilization. For example,
on a big.LITTLE system, if a task has a utilization above the capacity
of the LITTLE cluster, than its min_cluster_cap is 1024 (which is the capacity
of the big cluster where this task should be scheduled).

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>