libs/utils/analysis: move filters APIs into TASKs module
All the functions defined from libs/utils/filters.py, in the new refactoring
of the TraceAnalysis code, they better belong to the tasks_analysis module.
Let's move this code under this module as well as refactor a big the
DataFrame getter methods to return a more complete set of tables.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>