tests/eas/hmp_parity: don't clobber big tasks
When creating the tasks for the ForkMigration test the test framework was
creating the big and little tasks with the same prefix, which meant that
the loop that creates the little tasks overwrote the tasks created in the
previous big task loop.

Use different prefixes for big and little tasks so that they are all
created.  55eee3cb6f1f ("tests/eas/hmp_parity: Add descriptions for
tests") described the ForkMigration behaviour as it was, instead of the
spirit of the test: create as many threads as there are cores in the
system.  Fix the description of the test to what the test always meant
to do and now actually does.