libs/utils/energy: make energy meters API more generic
Currently available energy meters provide different formats for the results,
which makes the API not generic.

This makes the HWMon energy meters return results with similar format as the
ACME. A new attribute is introduced in the configuration dictionary called
"channel_map" that allows to map a user-defined label to the energy
meter-specific channel name. For example, in case of HWMon channel names are
the strings specified as "sites".

This way, when using the report() method, the user knows what is the label she
should use to measuree energy from a given channel and allows compatibility of
the code when using a different type of energy meter, by just changing the
channel mapping.

Signed-off-by: Michele Di Giorgio <michele.digiorgio@arm.com>