tests/eas/hmp_parity: refactor the OffloadMigrationAndIdlePull test
The Offload Migration test assumes that the early starters will run in
the big cpus until completion and then the "migrator" tasks are going to
be pulled from the littles.

Recent versions of EAS may not follow this pattern exactly, even though
their behaviour is still valid.  It is possible for a migrator task to
displace an early_starter task and run in a big cpu since it starts
running, forcing the early starter to run on a little cpu until a big
cpu is available.

The new tests now check that the big cpus were fully busy as long as
there are any big tasks still active and that all the tasks got to run
in the big cpus at some point.  This verifies the idle pull behaviour
but still allows for early starters to be displaced.