tests/eas/hmp_parity: cache the trappy.FTrace object in OffloadMigrationAndIdlePull
In the OffloadMigrationAndIdlePull test we call SchedMultiAssert
multiple times with the trace filename as a parameter, making
each SchedMultiAssert create a trappy object for itself.

Call trappy.FTrace from the test itself so that we can pass a reference
to it to all the SchedMultiAssert invocations, avoiding multiple copies
of effectively the same thing.