EventPlot: Add ability to plot simultaneous events
The timeline plot was designed to have one "key" running in one "lane".
Here you have the "queueing" key running in multiple lanes and this
confuses the filtering logic. I imagined the each event to be only
present in one lane at a time, much like a single PID can be only on a
single CPU.

This fix allows the EventPlot to plot concurrent events. Events which
occur simultaneously in multiple lanes