libs/wlgen/workload: allows different CGroups and CPUs for each run
In the current implementation, the CPUs and on CGroups where a workload
has to be executed at each run() call it's somehow defined either at
configuration time or at its first execution. This is due to the way we
add the taskset and/or cgroup_run_into prefixes.

This patch builds a new _command at each run() invocation which allows
each time to prefix the workload command with the proper taskset and/or
cgroup_run_into. This allows to use different CPUs and/or CGroups each time
we run a preconfigured workload.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>