libs/wlgen/workload: avoid special case for CGroups into execution
Currently we have a special case for execution of a workload within
a CGroup which makes the code odd with many if cases.

This patch simplifies the code by formatting the shutils based
command when we define the CGroup run-into command.
Moreover, all execution commands honour the as_root parameter specified
by the run call.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>