libs/utils/analysis: handle events with duplicate timestamps
It might happen that a trace has two (or more) events with the same timestamp.
Pandas does not allow reindexing of duplicate indexes.

For this reason, cluster active signal is built by joining dataframes instead of
creating the dataframe from a dictionary of series we perform a join between
dataframe of active signals.

Also, frequency residency computation does not perform reindexing anymore but
rather a join operation. As a consequence, because series data structures do not
support join, it is necessary to undo what 37a7192 does and keep `cluster_freqs`
as a dataframe instead.

Signed-off-by: Michele Di Giorgio <michele.digiorgio@arm.com>