tests/eas/hmp_parity: add some more interesting sched ftrace events
HMP parity tests run for only a few seconds.  We can collect more ftrace
events on each run as they should not use too much space and can help in
debugging issues.

sched_energy_diff, sched_load_avg_task and sched_load_avg_cpu are not in
mainline but they are part of the EAS patchset.  I'm including them
because it is expected that the test will be run on systems with the EAS
patches applied.