sched: make sched_migrate_task parse the normal trace from trace-cmd
The EAS patchset includes an enhancement of sched_migrate_task that
includes a enum field printed with __print_symbolic that only gets
parsed in the trace.txt and not in the trace.raw.txt.

Remove the instances of sched_migrate_task from the trace_empty file to
keep the tests testing a trace that is "empty", that is, no event in it
is parsed.