tasks_analysis: remove sched_overutilized from plot accounting
In plotTasks(), sched_overutilized must not be considered when deciding whether
to generate one of the three plots (utilization, residency, PELT).

Furthermore, sched_overutilized is only shown when either the user specifies it
as one of the signals to be plot or no signals are specified (so, by default all
three plots are generated with all signals).

Signed-off-by: Michele Di Giorgio <michele.digiorgio@arm.com>