libs/utils/trace: refactor computation of the plot window
Compute the plot window values in a separate private method and set the window
attribute of the trace object accordingly.

In case the user does specify any window, the upper bound is set to the maximum
duration of the trace, so that we do not have to deal with a NoneType value
later in methods that take the plot window into account.

Signed-off-by: Michele Di Giorgio <michele.digiorgio@arm.com>