tests/eas/acceptance: Increase required busy time for idle pull test
90% is too forgiving for EXPECTED_BIG_BUSY_TIME_PCT, it allows traces to
pass the "Bug cpus are fully loaded" test even when 100% tasks have run
on LITTLE CPUs while big CPUs are idle for significant periods.

Since traces contain a perdiod of inactivity at the beinning, to require
99% busy time it's necessary to adjust the trace window over which
assertions apply. Instead of starting at t=0 we use
SchedMultiAssert.getStartTime to find the time at which the first task
started, and use that as the start of our window.