tests/eas/acceptance: increase accuracy of little cpus test in the Offload Migration testcase
The window for test_little_cpus_run_tasks() was loosely defined as just
the migrator delay, which doesn't take into account the time it takes
for rt-app to spawn and create its tasks.  After
fa4bce7359ec ("tests/eas/acceptance: Increase required busy time for
idle pull test"), the test is more strict, so increase the accuracy of
the window to meet the new restrictions.