tests/eas/acceptance: enable sched_is_big_little sysctl
Where this sysctl is present, if it isn't enabled the EAS wake path
views the CPUs as a flat list, so that LITTLE CPUs are not necessarily
preferred over big when selecting CPUs for a low-load task. Set it, so
that the WakeMigration and other tests can pass.

Since this flag isn't used by the mainline-integration version of EAS,
we only attempt to set it if the procfs file is present.