utils/env: add support for ANDROID_HOME and CATAPULT_HOME
These environment variables are required by devlib and some notebooks
to properly access and trace an Android device. This path add the required
support to verify that these env variables are defined when an Android
target is in use as well as the required support to override these
variables via the target configuration.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>