analysis/perf: add new module to support RTApp performance analysis
This is a re factored version of the perf_analysis which is integrated
with the "data_frame" and "analysis" namespaces.

NOTE: Currently, this patch makes the Performance Analysis
depending on the Trace module, which currently requires a trace to
be properly initialized. Thus, it is not more possible to have a performance
analysis if a trace is not collected.

A following patch will allow to initialize a Trace object even when a trace
is not available.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>