tests/eas/acceptance: Reduce expected LITTLE busy time for offload test
Where the 'migrator' tasks in the OffloadMigrationAndIdlePull are
initially placed on the same CPU, the busy time of the LITTLE CPUs that
are exercised can be below 99% during the period where big CPUs are
occupied. This shouldn't be considered a failure.

In my experiments 97% appeared to be a low enough value to prevent these
failures. In fact I observed no failures with the value at 98% either,
but it seems wise to allow a little margin for error.