libs/wlgen/rta: Limit task name to 15 characters
rt-app uses pthread_setname_np(3) to set the names of tasks on the
target [1]. This interface limits thread name length to 16 characters
including the null '\0'.

Although rt-app prints a message on the target, the workload execution
goes ahead when the provided task names are too long, and the tasks are
named "rt-app". This means they are not found when later examining the
trace.

Therefore, raise an exception on the host to prevent this situation
arising.

[1] https://github.com/scheduler-tools/rt-app/blob/dcaf5a04aa63b69d8b5e174d10a9e459f88772cf/src/rt-app.c#L152