libs/utils/env: Explicitly check 'host' setting is present for Linux
For Linux targets you need a 'host' parameter to set up the SSH
connection. If it's missing, the error appears at the end of a deep dark
call stack.

Add an explicit check before constructing the LinuxTarget so it's
more obvious what you did wrong when this param is missing.