executor: Add flag to freeze userspace while running experiments
This is to reduce system noise.

Note that if the cgroups devlib module is not present, this flag will
case an _error_. If the cgroups module is present but the freezer
cgroups controller is not preset on the target, this flag will just
cause a _warning_. This is because it's judged that targets without the
freezer controller probably have quiet enough userspace that they don't
need it.