tests/eas: Add test for frequency invariant load tracking
This test runs the same workload on the most capable CPU on the system
at a cross section of available frequencies. The trace is then examined
to find the average activation length of the workload, which is combined
with the known period and frequency to estimate an expected mean value
for util_avg for each frequency. The util_avg value is extract from
scheduler trace events and its mean is compared with the expected
value (ignoring the first 300ms so that the signal can stabilize). The
test fails if the observed mean is beyond a certain error margin from
the expected one. load_avg is then similarly compared with the expected
util_avg mean, under the assumption that load_avg should equal util_avg
when system load is light.

The test is many helper methods, so that in the case of test failures
the logic used by the test can be re-used to examine the _cause_ of the
failure. A canned notebook is provided with basic plots showing test
results in a more useful format than the command-line test run.