sched: add first simple PELT simulator
This patch provide a first implementation of a simple PELT simulator.
Despite being of a bare simplicity it's still a useful tool to show and
study how PELT describes a task depending on its profile (period and
duty-cycle) as well as some specific PELT attributes (half-life, initial
values).
The current PELT model can be easily extended to evaluate experimental
features such as (for example) the capping of the decay value.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Reviewed-by: KP Singh <kpsingh@google.com>