LisaTest: Add methods to override for test configuration
These methods allow users of LisaTest to provide a `test_conf` member
which configures the test environment (TestEnv) and an
`experiments_conf` member to configure the experiments to set up the
experiments (Executor).

Users can override _getExperimentsConf with a method that generates the
Executor experiments configuration as a function of target information -
this is called _aftere_ the TestEnv is initialised and it is passed as a parameter.