lisa_shell: update test command
Currently the test command does not allow to use the bash completion to
defined the test to execute. This is quite annoying since we are masking
unecessary a really useful shell functionality

This patch fixes this as well as it enforce the usage of the locally defined
logging configuration file as well as ensure to report in output all the
log statements. This allows to monitor tests progress while the test is
in execution and avoid to confuse the user which can otherwise think that
the test is not running.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>